refactor: rename and reorganize filesystem utility functions
- Updated import paths for filesystem utility functions to improve clarity and consistency.
- Replaced `getFileSystemScs` with `getFileSystemStorageClasses` for better naming.
- Removed obsolete filesystem utility file.
